rrr: convert base/callback_wrapper to module (rrr.callback_wrapper)
callback_wrapper.cpp is the rrr.callback_wrapper module interface
unit; .hpp deleted. Pure templated helper: rrr::detail::CallbackWrapper<Sig>
wrapping rusty::Arc<rusty::Function<Sig>> to make move-only Functions
copyable for places like channel callbacks and alock callbacks.

Includers (2) updated: rpc/channel.hpp, misc/alock.hpp. 26 modules.
